Move androidTest files
